LayoutTests/fast/profiler/heavy-view.html

<html>
<head>
<script src="resources/profiler-test-JS-resources.js"></script>
<script>
if (window.layoutTestController)
    layoutTestController.dumpAsText();

console.profile("Heavy View Test");

function a()
{
    b();
    foo();
    d();
    d();
    g();
    console.profileEnd();
    printHeavyProfilesDataWithoutTime();
}

function b()
{
    foo();
    foo();
}

function c()
{
    foo();
}

function d()
{
    e();
}

function e()
{
    foo();
}

function g()
{
    b();
}

var firstTime = true;
function foo()
{
    if (firstTime) {
        firstTime = false;
        c();
    }
}

</script>
</head>

<body onload="a()">
This page has several functions calling each other in a way that exposes interesting
edge cases related to the Heavy view in the profiler.
<br>
<br>
To run this test manually, load it in the browser then load the WebInspector and look at
the profile in Heavy view.
<div id="output"></div>
</body>
</html>
